+++ Observer/CreditmemoRefundObserver.php
@@ -22,6 +22,8 @@ use SDM\Altapay\Helper\Config as storeConfig;
 use SDM\Altapay\Model\Handler\OrderLinesHandler;
 use SDM\Altapay\Model\Handler\PriceHandler;
 use SDM\Altapay\Model\Handler\DiscountHandler;
+use Magento\Framework\Math\Random;
+use SDM\Altapay\Model\ReconciliationIdentifierFactory;
 
 /**
  * Class CreditmemoRefundObserver
@@ -62,18 +64,28 @@ class CreditmemoRefundObserver implements ObserverInterface
      * @var DiscountHandler
      */
     private $discountHandler;
-
+    /**
+     * @var ReconciliationIdentifierFactory
+     */
+    private $reconciliation;
+    /**
+     * @var Random
+     */
+    private $random;
+    
     /**
      * CreditmemoRefundObserver constructor.
      *
-     * @param SystemConfig      $systemConfig
-     * @param Logger            $altapayLogger
-     * @param Order             $order
-     * @param Data              $helper
-     * @param storeConfig       $storeConfig
-     * @param OrderLinesHandler $orderLines
-     * @param PriceHandler      $priceHandler
-     * @param DiscountHandler   $discountHandler
+     * @param SystemConfig                    $systemConfig
+     * @param Logger                          $altapayLogger
+     * @param Order                           $order
+     * @param Data                            $helper
+     * @param storeConfig                     $storeConfig
+     * @param OrderLinesHandler               $orderLines
+     * @param PriceHandler                    $priceHandler
+     * @param DiscountHandler                 $discountHandler
+     * @param ReconciliationIdentifierFactory $reconciliation
+     * @param Random                          $random
      */
     public function __construct(
         SystemConfig $systemConfig,
@@ -83,7 +95,9 @@ class CreditmemoRefundObserver implements ObserverInterface
         storeConfig $storeConfig,
         OrderLinesHandler $orderLines,
         PriceHandler $priceHandler,
-        DiscountHandler $discountHandler
+        DiscountHandler $discountHandler,
+        ReconciliationIdentifierFactory $reconciliation,
+        Random $random
     ) {
         $this->systemConfig    = $systemConfig;
         $this->altapayLogger   = $altapayLogger;
@@ -93,6 +107,8 @@ class CreditmemoRefundObserver implements ObserverInterface
         $this->orderLines      = $orderLines;
         $this->priceHandler    = $priceHandler;
         $this->discountHandler = $discountHandler;
+        $this->reconciliation  = $reconciliation;
+        $this->random          = $random;
     }
 
     /**
@@ -250,16 +266,13 @@ class CreditmemoRefundObserver implements ObserverInterface
     private function sendRefundRequest($memo, $orderLines, $orderObject, $payment, $storeCode)
     {
         $refund = new RefundCapturedReservation($this->systemConfig->getAuth($storeCode));
-        $reconciliationIdentifier  = $payment->getAdditionalInformation('altapay_reconciliation');
+        $reconciliationIdentifier  = $this->random->getUniqueHash();
         if ($memo->getTransactionId()) {
             $refund->setTransaction($payment->getLastTransId());
         }
         $refund->setAmount((float)number_format($memo->getGrandTotal(), 2, '.', ''));
         $refund->setOrderLines($orderLines);
-
-        if(!empty($reconciliationIdentifier)){
-            $refund->setReconciliationIdentifier($reconciliationIdentifier);
-        }
+        $refund->setReconciliationIdentifier($reconciliationIdentifier);
         try {
             $refund->call();
         } catch (ResponseHeaderException $e) {
@@ -268,6 +281,14 @@ class CreditmemoRefundObserver implements ObserverInterface
         } catch (\Exception $e) {
             $this->altapayLogger->addCriticalLog('Exception: ' , $e->getMessage());
         }
+
+        $model = $this->reconciliation->create();
+        $model->addData([
+            "order_id"      => $memo->getOrder()->getIncrementId(),
+            "identifier"    => $reconciliationIdentifier,
+            "type"          => 'refunded'
+        ]);
+        $model->save();
 
         $rawResponse = $refund->getRawResponse();
         $body        = $rawResponse->getBody();